Update README with travis/version links and more docs with actual usage
