+++ Gemfile
@@ -2,7 +2,3 @@ source "https://rubygems.org"
 
 # Specify your gem's dependencies in trailblazer-test.gemspec
 gemspec
-
-group :test do
-  gem "minitest-line"
-end
