+++ test/test_helper.rb
@@ -1,3 +1,10 @@
+require "bundler/setup"
+require "simplecov"
+SimpleCov.start do
+  add_group "Trailblazer-Test", "lib"
+  add_group "Tests", "test"
+end
+
 $LOAD_PATH.unshift File.expand_path("../lib", __dir__)
 require "trailblazer/test"
 
@@ -46,6 +53,44 @@ Minitest::Spec.class_eval do
 
     def failure?
       !@success
+    end
+  end
+
+  class Create
+    def self.call(params:)
+      if params[:band] == "Rancid"
+        model = Struct.new(:title, :band).new(params[:title].strip, params[:band])
+        Result.new(true, model, nil)
+      else
+
+        Result.new(false, nil, Errors.new(band: ["must be Rancid"]))
+      end
+    end
+  end
+
+  class Update
+    def self.call(params:, current_user:)
+      return Result.new(false, nil, Errors.new(nil, Policy.new(false))) if current_user.name != "allowed"
+
+      if params[:band] == "Rancid"
+        model = Struct.new(:title, :band).new(params[:title].strip, params[:band])
+        Result.new(true, model, nil)
+      else
+
+        Result.new(false, nil, Errors.new(band: ["must be Rancid"]))
+      end
+    end
+  end
+
+  class CreateNestedParams
+    def self.call(params:)
+      if params[:form][:band] == "Rancid"
+        model = Struct.new(:title, :band).new(params[:form][:title].strip, params[:form][:band])
+        Result.new(true, model, nil)
+      else
+
+        Result.new(false, nil, Errors.new(band: ["must be Rancid"]))
+      end
     end
   end
 end
